requirements.txt
requests>=2.28.0
websockets>=11.0.0
urllib3>=2.0.0
python-dotenv>=1.0.0
